+++ src/rogallo/screens/main/gemini.py
@@ -177,7 +177,7 @@ async def _handle_response(
     if is_displayable_mime_type(response.mime_type):
         owner.post_message(
             OpenDocument(
-                document=cache.add_document(
+                cache.add_document(
                     Document(
                         location=uri,
                         original_location=request.location,
@@ -187,9 +187,10 @@ async def _handle_response(
                         verification_method=response.verification_method,
                         server_certificate=response.server_cert,
                         avoid_cache=response.client_cert_used,
+                        avoid_history=request.avoid_history,
                     )
                 ),
-                original_request=request,
+                from_history=request.from_history,
             )
         )
     else:
@@ -221,12 +222,17 @@ async def handle_gemini_request(
 
     # If a cached copy of the document exists and the request allows it,
     # use that instead of making a network request.
-    if request.allow_cached and (cached_document := cache.get_document(uri)):
-        owner.post_message(
-            OpenDocument(
-                document=cached_document,
-                original_request=request,
+    if (
+        request.allow_cached
+        and (
+            cached_document := cache.get_document(
+                uri, avoid_history=request.avoid_history
             )
+        )
+        is not None
+    ):
+        owner.post_message(
+            OpenDocument(document=cached_document, from_history=request.from_history)
         )
         return