+++ src/rogallo/screens/main/nex.py
@@ -33,12 +33,11 @@ async def handle_nex_request(
     assert isinstance(uri, NexURI)
 
     # Check the cache first.
-    if request.allow_cached and (cached_document := cache.get_document(uri)):
+    if request.allow_cached and (
+        cached_document := cache.get_document(uri, avoid_history=request.avoid_history)
+    ):
         owner.post_message(
-            OpenDocument(
-                document=cached_document,
-                original_request=request,
-            )
+            OpenDocument(cached_document, from_history=request.from_history)
         )
         return
 
@@ -66,16 +65,17 @@ async def handle_nex_request(
     if is_displayable_mime_type(response.mime_type):
         owner.post_message(
             OpenDocument(
-                document=cache.add_document(
+                cache.add_document(
                     Document(
                         location=uri,
                         original_location=uri,
                         content=response.text,
                         mime_type=response.mime_type,
                         avoid_cache=False,
+                        avoid_history=request.avoid_history,
                     )
                 ),
-                original_request=request,
+                from_history=request.from_history,
             )
         )
     else: